% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundaryBuilder.R
\name{boundaryBuilder}
\alias{boundaryBuilder}
\title{Create the boundary file to the conservation management plannning software}
\usage{
boundaryBuilder(
  file,
  distance_limit = .Machine$integer.max,
  connection_limit = .Machine$integer.max,
  export_csv = FALSE
)
}
\arguments{
\item{file}{Object of class [data.frame()] or [SpatialLinesDataFrame()] or [SpatialPolygonsDataFrame()] that specifies
the relation between planning (river PU's). This file must contain the following fields:
\describe{
   \item{`GridID`}{`integer` unique identifier for each planning unit.}
   \item{`NextDownID`}{`integer` unique identifier for the planning unit immediately below. Set -1
   there isn't one below.}
   \item{`Length`}{`numeric` value of distance between the `GridID` and `NextDownID`.}
   }}

\item{distance_limit}{distance limit to stop to building the relation between units.}

\item{connection_limit}{number of connections limit to stop to building the relation between units.}

\item{export_csv}{is TRUE to export a .csv file with boundary connections.}
}
\value{
A [data.frame()].
}
\description{
Return a data.frame
generated by information about the boudary data from river data.
}
\examples{
\donttest{
boundaryBuilder("inst/extdata/Douro2.csv", distance_limit = 100, connection_limit = 20, export_csv = FALSE)
}

}
