% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPuBarriers.R
\name{createPuBarriers}
\alias{createPuBarriers}
\title{Create River Network Planning Units with Barriers}
\usage{
createPuBarriers(
  reaches_sf,
  barriers_sf,
  threshold_distance = 100,
  near_distance_barriers = 10,
  id = "GridID",
  nxt = "NextDownID"
)
}
\arguments{
\item{reaches_sf}{An `sf` object representing the river network. It must contain the columns `HYRIV_ID` (unique 
identifier for each segment) and `NEXT_DOWN` (identifier for the downstream segment).}

\item{barriers_sf}{An `sf` object representing the barriers. Points in this dataset will be projected 
onto the river network.}

\item{threshold_distance}{Numeric. The maximum distance allowed for a barrier to be considered 
"close enough" to a river segment.}

\item{near_distance_barriers}{Numeric. The minimum distance between projected barrier points. 
Barriers closer than this distance will be aggregated.}
}
\value{
A list with two components:
\describe{
  \item{`divided_reaches`}{An updated `sf` object representing the river network with split segments.}
  \item{`projected_points`}{An `sf` object containing the projected locations of the barriers.}
  \item{`final_reaches`}{An `sf` object containing the final reaches.}
}
}
\description{
This function modifies a river network's spatial structure (`reaches_sf`) by incorporating the 
effects of barriers (`barriers_sf`). The function projects barriers onto the nearest river segments, adjusts 
the connectivity relationships (e.g., `NEXT_DOWN`), and returns an updated river network.
}
\examples{
\donttest{
 # Example usage
 result <- createPuBarriers(reaches_sf, barriers_sf, threshold_distance = 100, near_distance_barriers = 10)
 updated_reaches <- result$updated_reaches
 projected_points <- result$projected_points
}

}
